+++ releasenotes.md
@@ -2,6 +2,21 @@
 
 ## Release Notes
 
+### v1.2.6
+  * Add instructions to create cowsayings directory structure in the manifests
+    and classes quest instead of relying on it being pre-created in the build.
+
+### v1.2.5
+  * Content tested for compatibility with puppet-2016.2.1-learning-5.6 VM build.
+  * Updated [pltraining-bootstrap](https://github.com/puppetlabs/pltraining-bootstrap) module turn off default line-numbering in vim.
+  * Updated [pltraining-learning](https://github.com/puppetlabs/pltraining-learning) module to create module structure directories for cowsay module.
+
+### v1.2.4
+  * Content tested for compatibility with puppet-2016.2.1-learning-5.5 VM build.
+  * Updated CSS styling.
+  * Minor changes to tests to be compatible with RESTful quest tool version. 
+  * Minor content fixes.
+
 ### v1.2.3
   * Content tested for compatibility with puppet-2016.2.0-learning-5.4 VM build.
   * Added privacy a link to Puppet's privacy policy.